Major changes to key sig verification
Added verification for key signatures.

Any commands coming in are accompanied by a signature. The signature is used alongside the public key to verify commands are issued from the authorizing private key. In the future, each beacon will be generated with its own public/private key and commands outgoing will be signed through a simple lookup