Updating readability
The controller had a bunch of manually retyped blocks of text. These should have been const's from the start and they are now. Just edit the block once to affect how they all print it. Same with the help menu.

I also identified a switch case with no default scenario when asking if you want to use the profile.